Bring sitemap-all at the root
