--- apps/frontpage/app/sitemap-all/sitemap.xml
@@ -1,11 +1,0 @@
-<sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-  <sitemap>
-    <loc>https://storybook.js.org/sitemap-all/addons/sitemap.xml</loc>
-  </sitemap>
-  <sitemap>
-    <loc>https://storybook.js.org/sitemap-all/docs/sitemap.xml</loc>
-  </sitemap>
-  <sitemap>
-    <loc>https://storybook.js.org/sitemap-all/recipes/sitemap.xml</loc>
-  </sitemap>
-</sitemapindex>
